# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Finance Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: getAccounts
      tags: [accounts]
      summary: Get all accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountInput"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    put:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountInput"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
  /budgets:
    get:
      operationId: getBudgets
      tags: [budgets]
      summary: Get budgets for a month
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of budgets with spending
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetWithSpending"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create or update budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetInput"
      responses:
        "201":
          description: Created budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /budgets/{id}:
    put:
      operationId: updateBudget
      tags: [budgets]
      summary: Update budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetInput"
      responses:
        "200":
          description: Updated budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: Get dashboard summary
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"
  /analytics/by-category:
    get:
      operationId: getAnalyticsByCategory
      tags: [analytics]
      summary: Get spending by category
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
      responses:
        "200":
          description: Spending by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryAnalytics"
  /analytics/monthly-trend:
    get:
      operationId: getMonthlyTrend
      tags: [analytics]
      summary: Get monthly income/expense trend
      parameters:
        - name: months
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Monthly trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyTrend"
...
  /analytics/export:
    get:
      operationId: exportTransactions
      tags: [analytics]
      summary: Export transactions as CSV
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: format
          in: query
          schema:
            type: string
            enum: [csv]
            default: csv
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
...
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [cash, card, wallet, savings, investment]
        currency:
          type: string
        initialBalance:
          type: number
        currentBalance:
          type: number
        color:
          type: string
        icon:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, type, currency, initialBalance, currentBalance, createdAt]
    CreateAccountInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [cash, card, wallet, savings, investment]
        currency:
          type: string
        initialBalance:
          type: number
        color:
          type: string
        icon:
          type: string
      required: [name, type, currency, initialBalance]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [income, expense, both]
        color:
          type: string
        icon:
          type: string
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, type, color, icon, isDefault, createdAt]
...
    Budget:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        amount:
          type: number
        year:
          type: integer
        month:
          type: integer
        createdAt:
          type: string
          format: date-time
        category:
          $ref: "#/components/schemas/Category"
      required: [id, categoryId, amount, year, month, createdAt]
    BudgetWithSpending:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        amount:
          type: number
        spent:
          type: number
        year:
          type: integer
        month:
          type: integer
        createdAt:
          type: string
          format: date-time
        category:
          $ref: "#/components/schemas/Category"
        previousMonthSpent:
          type: number
      required: [id, categoryId, amount, spent, year, month, createdAt]
    CreateBudgetInput:
      type: object
      properties:
        categoryId:
          type: integer
        amount:
          type: number
        year:
          type: integer
        month:
          type: integer
      required: [categoryId, amount, year, month]
...
    AnalyticsSummary:
      type: object
      properties:
        totalBalance:
          type: number
        monthIncome:
          type: number
        monthExpense:
          type: number
        monthNet:
          type: number
        previousMonthIncome:
          type: number
        previousMonthExpense:
          type: number
        accountBalances:
          type: array
          items:
            $ref: "#/components/schemas/AccountBalance"
        topExpenseCategories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryAnalytics"
        projectedBalance:
          type: number
      required: [totalBalance, monthIncome, monthExpense, monthNet, accountBalances, topExpenseCategories]
    AccountBalance:
      type: object
      properties:
        accountId:
          type: integer
        accountName:
          type: string
        balance:
          type: number
        currency:
          type: string
      required: [accountId, accountName, balance, currency]
    CategoryAnalytics:
      type: object
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryColor:
          type: string
        categoryIcon:
          type: string
        total:
          type: number
        count:
          type: integer
        percentage:
          type: number
      required: [categoryId, categoryName, categoryColor, categoryIcon, total, count, percentage]
    MonthlyTrend:
      type: object
      properties:
        year:
          type: integer
        month:
          type: integer
        income:
          type: number
        expense:
          type: number
        net:
          type: number
      required: [year, month, income, expense, net]
...