# lib/api-zod/src/generated/types/accountBalance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface AccountBalance {
  accountId: number;
  accountName: string;
  balance: number;
  currency: string;
}