# lib/api-zod/src/generated/types/budgetWithSpending.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";

export interface BudgetWithSpending {
  id: number;
  categoryId: number;
  amount: number;
  spent: number;
  year: number;
  month: number;
  createdAt: Date;
  category?: Category;
  previousMonthSpent?: number;
}