# lib/db/src/schema/accounts.ts
import { pgTable, serial, text, numeric, timestamp, pgEnum } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const accountTypeEnum = pgEnum("account_type", ["cash", "card", "wallet", "savings", "investment"]);

export const accountsTable = pgTable("accounts", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  type: accountTypeEnum("type").notNull().default("card"),
  currency: text("currency").notNull().default("RUB"),
  initialBalance: numeric("initial_balance", { precision: 15, scale: 2 }).notNull().default("0"),
  currentBalance: numeric("current_balance", { precision: 15, scale: 2 }).notNull().default("0"),
  color: text("color").notNull().default("#6366f1"),
  icon: text("icon").notNull().default("account_balance_wallet"),
  createdAt: timestamp("created_at").notNull().defaultNow(),
});

export const insertAccountSchema = createInsertSchema(accountsTable).omit({ id: true, createdAt: true, currentBalance: true });
export type InsertAccount = z.infer<typeof insertAccountSchema>;
export type Account = typeof accountsTable.$inferSelect;